Hide sa test for now
The bulk of the sa stuff has been implemented. There is a fundamental flaw with how the primitive functions are handled.

Moving forward this branch will be merged, a new branch will be created to resolve the primitives issue, then another branch will be created to resolve any final issues.